chore: move quizz domain to account
- for now. this is one of the last field that will not be migrated to kratos.

it should eventually be it's own collection but this can be done when a migration out of the main codebase is done with an apollo federation
