+++ src/graphql/root/query/me.ts
@@ -4,7 +4,10 @@ import GraphQLUser from "@graphql/types/object/graphql-user"
 
 const MeQuery = GT.Field({
   type: GraphQLUser,
-  resolve: async (_, __, { domainUser }) => domainUser,
+  resolve: async (_, __, { domainUser, domainAccount }) => ({
+    ...domainUser,
+    quizQuestions: domainAccount?.quizQuestions,
+  }),
 })
 
 export default MeQuery
